+++ cdm_reader_mapper/mdf_reader/__init__.py
@@ -2,9 +2,9 @@
 
 from __future__ import annotations
 
-from .data_models import code_tables  # noqa
-from .data_models import schemas  # noqa
+from .code_tables import code_tables  # noqa
 from .get_sections import get_sections  # noqa
 from .import_data import import_data  # noqa
 from .read import read  # noqa
 from .read_sections import read_sections  # noqa
+from .schema import schemas  # noqa